+++ packages/core/src/manglers/__tests__/css-variables.test.ts
@@ -128,6 +128,10 @@ suite("CSS Variable Mangler", function() {
             input: "div { color: var(--foo); } p { font: var(--bar); }",
             expected: "div { color: var(--a); } p { font: var(--b); }",
           },
+          {
+            input: "div { margin: 0 var(--foo); }",
+            expected: "div { margin: 0 var(--a); }",
+          },
         ],
       },
       {
@@ -341,6 +345,20 @@ suite("CSS Variable Mangler", function() {
             .filter((testCase) => !/(\s|^)(style)=/.test(testCase.input)),
         ]
         .flatMap(varyHtmlQuotes)
+        .flatMap(embedAttributesInTags),
+      },
+      {
+        name: "multi-value CSS declaration",
+        cases: [
+          {
+            input: "style=\"margin: 0 var(--foobar);\"",
+            expected: "style=\"margin: 0 var(--a);\"",
+          },
+          {
+            input: "style=\"margin: 1em var(--foo) 1px var(--bar);\"",
+            expected: "style=\"margin: 1em var(--a) 1px var(--b);\"",
+          },
+        ]
         .flatMap(embedAttributesInTags),
       },
       {